Modify models db

+++ SferumNet.DbModels/Common/Entity.cs
@@ -4,5 +4,5 @@ namespace SferumNet.DbModels.Common;
 
 public class Entity
 {
-    [Key] public int Id { get; set; }
+    [Key] public long Id { get; set; }
 }
